+++ tests/helpers.py
@@ -459,6 +459,7 @@ TEST_MASK_URL = "https://raw.githubusercontent.com/scaleapi/nucleus-python-clien
 
 this_dir = os.path.dirname(os.path.realpath(__file__))
 TEST_LOCAL_MASK_URL = os.path.join(this_dir, "testdata/000000000285.png")
+TEST_LOCAL_TESTDIR = os.path.join(this_dir, "testdata/testdir")
 
 
 NUM_VALID_SEGMENTATIONS_IN_MAIN_DATASET = len(TEST_DATASET_ITEMS)